Several usability improvements with examples such as specifying the application and demand profile using a file, delegating messaging of client responses to gigapaxos, etc. These are in reconfiguration.examples and gigapaxos.examples. The READMEs have been updated accordingly.
In the previous commit, an example of batched name creation was added
in reconfiguration.testing.ReconfigurableClientCreateTester. This
speeds up name creation by an order of magnitude when all of the names
are being created on the same set of active replicas (but with
possibly different initial states if needed).

+++ src/edu/umass/cs/gigapaxos/paxospackets/AcceptReplyPacket.java
@@ -46,7 +46,7 @@ public class AcceptReplyPacket extends PaxosPacket {
 	 */
 	public final int maxCheckpointedSlot;
 
-	private int requestID = 0; // used only for debugging
+	private long requestID = 0; // used only for debugging
 
 	public AcceptReplyPacket(int nodeID, Ballot ballot, int slotNumber,
 			int maxCheckpointedSlot, AcceptReplyPacket ar) {
@@ -65,7 +65,7 @@ public class AcceptReplyPacket extends PaxosPacket {
 
 	// used only for instrumentation
 	public AcceptReplyPacket(int nodeID, Ballot ballot, int slotNumber,
-			int maxCheckpointedSlot, int requestID) {
+			int maxCheckpointedSlot, long requestID) {
 		this(nodeID, ballot, slotNumber, maxCheckpointedSlot);
 		this.setRequestID(requestID);
 	}
@@ -84,11 +84,11 @@ public class AcceptReplyPacket extends PaxosPacket {
 				.toString());
 	}
 
-	public void setRequestID(int id) {
+	public void setRequestID(long id) {
 		this.requestID = id;
 	}
 
-	public int getRequestID() {
+	public long getRequestID() {
 		return this.requestID;
 	}
